refactor: 공통 badge 컴포넌트 파일 정리
component/ui/badge 디렉토리 생성하고 badge 컴포넌트들 이동
